requirements.txt
numpy==1.21.0
pandas==1.3.0
tensorflow==2.8.0
yfinance==0.2.36
scikit-learn==0.24.2
Pillow==8.3.0
matplotlib==3.4.0
seaborn==0.11.0
ta==0.10.2  # Alternative to TA-Lib